reuse callback, allow user to acces all stable_baselines3 agents via config
